refactor: move out code from general.py
